app/requirements.txt
certifi==2021.10.8
elastic-transport==8.1.2
elasticsearch==8.2.0
psycopg2-binary==2.9.3
pydantic==1.9.0
python-stdnum==1.17
typing_extensions==4.2.0
urllib3==1.26.9
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
decorator==5.1.1
Django==3.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-split-settings==1.1.0
executing==0.8.3
flake8==4.0.0
gunicorn==20.1.0
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
